requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/LLM dependencies
openai>=1.12.0,<2.0.0
transformers==4.35.2  # Downgraded for compatibility
huggingface_hub==0.19.4  # Compatible version

# Vector database and embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Database dependencies
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Additional scraping tools
selenium==4.16.0
playwright==1.40.0
fake-useragent==1.4.0

# Data processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# HTTP and utilities
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic-settings==2.1.0
